+++ KalikoCMS.Admin/Admin/Content/PropertyType/CollectionPropertyEditor.js
@@ -1,5 +1,5 @@
 ﻿(function (propertyEditor) {
-    propertyEditor.collection = {
+    propertyEditor.collection = propertyEditor.collection || {
         openDialog: function (container, valueField, className) {
             var callback = function (val, exerpt) {
                 var list = $(container).find("ul.sortable-collection");
@@ -40,8 +40,8 @@
 
 
 (function (dialogs) {
-    dialogs.openEditCollectionPropertyDialog = function(className, value) {
-        parent.openModal('Content/Dialogs/EditCollectionPropertyDialog.aspx?propertyType=' + className + '&value=' + escape(value), 780, 460);
+    dialogs.openEditCollectionPropertyDialog = dialogs.openEditCollectionPropertyDialog || function(className, value) {
+        parent.openModalViaPost('Content/Dialogs/EditCollectionPropertyDialog.aspx', { propertyType: className, value: value }, 780, 460);
         return false;
     };
 })(top.propertyEditor.dialogs || (top.propertyEditor.dialogs = {}));
@@ -55,7 +55,7 @@ $(document).ready(function () {
 
     function sortUpdate(event, ui) {
         var className = $(this).attr("data-type");
-        var value = '{"$types":{"KalikoCMS.PropertyType.CollectionProperty`1[[' + className +']], KalikoCMS.Engine":"1"},"$type":"1", "Items": [';
+        var value = '{"$type":"KalikoCMS.PropertyType.CollectionProperty`1[[' + className +']], KalikoCMS.Engine", "Items": [';
 
         $(this).children("li").each(function () {
             value += $(this).attr("data-value") + ",";