+++ KalikoCMS.Engine/Search/SearchQuery.cs
@@ -24,11 +24,13 @@ namespace KalikoCMS.Search {
             ReturnFromPosition = 0;
             NumberOfHitsToReturn = 100;
             MetaData = new string[] {};
+            InFields = new string[] {};
         }
 
         public int ReturnFromPosition { get; set; }
         public int NumberOfHitsToReturn { get; set; }
         public string SearchString { get; set; }
         public string[] MetaData { get; set; }
+        public string[] InFields { get; set; }
     }
 }